requirements.txt
streamlit>=1.28.0
openai>=1.3.0
anthropic>=0.7.0
pandas>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
pymongo>=4.6.0
dnspython>=2.4.0
lexicalrichness>=0.5.0
nltk>=3.8.0
textstat>=0.7.0
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.0
scipy>=1.10.0
